+++ packages/anthropic/src/tool/computer_20241022.ts
@@ -1,5 +1,5 @@
 import {
-  createProviderDefinedToolFactory,
+  createProviderToolFactory,
   lazySchema,
   zodSchema,
 } from '@ai-sdk/provider-utils';
@@ -26,7 +26,7 @@ const computer_20241022InputSchema = lazySchema(() =>
   ),
 );
 
-export const computer_20241022 = createProviderDefinedToolFactory<
+export const computer_20241022 = createProviderToolFactory<
   {
     /**
      * The action to perform. The available actions are: