+++ packages/provider-utils/src/index.ts
@@ -27,11 +27,11 @@ export { parseJsonEventStream } from './parse-json-event-stream';
 export { parseProviderOptions } from './parse-provider-options';
 export * from './post-to-api';
 export {
-  createProviderDefinedToolFactory,
-  createProviderDefinedToolFactoryWithOutputSchema,
-  type ProviderDefinedToolFactory,
-  type ProviderDefinedToolFactoryWithOutputSchema,
-} from './provider-defined-tool-factory';
+  createProviderToolFactory,
+  createProviderToolFactoryWithOutputSchema,
+  type ProviderToolFactory,
+  type ProviderToolFactoryWithOutputSchema,
+} from './provider-tool-factory';
 export * from './remove-undefined-entries';
 export * from './resolve';
 export * from './response-handler';