+++ packages/baseten/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/baseten
 
+## 1.0.0-beta.43
+
+### Patch Changes
+
+- Updated dependencies [544d4e8]
+  - @ai-sdk/openai-compatible@2.0.0-beta.41
+  - @ai-sdk/provider-utils@4.0.0-beta.40
+  - @ai-sdk/provider@3.0.0-beta.22
+
 ## 1.0.0-beta.42
 
 ### Patch Changes