+++ packages/black-forest-labs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/black-forest-labs
 
+## 1.0.0-beta.11
+
+### Patch Changes
+
+- Updated dependencies [544d4e8]
+  - @ai-sdk/provider-utils@4.0.0-beta.40
+  - @ai-sdk/provider@3.0.0-beta.22
+
 ## 1.0.0-beta.10
 
 ### Patch Changes