+++ packages/huggingface/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ai-sdk/huggingface
 
+## 1.0.0-beta.37
+
+### Patch Changes
+
+- 544d4e8: chore(specification): rename v3 provider defined tool to provider tool
+- Updated dependencies [544d4e8]
+  - @ai-sdk/openai-compatible@2.0.0-beta.41
+  - @ai-sdk/provider-utils@4.0.0-beta.40
+  - @ai-sdk/provider@3.0.0-beta.22
+
 ## 1.0.0-beta.36
 
 ### Patch Changes