+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/vue
 
+## 3.0.0-beta.124
+
+### Patch Changes
+
+- Updated dependencies [544d4e8]
+  - @ai-sdk/provider-utils@4.0.0-beta.40
+  - ai@6.0.0-beta.124
+
 ## 3.0.0-beta.123
 
 ### Patch Changes